Calendar with Events
Made a simple no UI calendar that gets your planned events from the google calendar. Does avoid anything past a month. Birthdays that repeat every year are also considered as events and are useless for this. So, just filtered out events that are not due in less than a month